requirements.txt
pycryptodome
numpy
PyWavelets
pydub